+++ src/Node.php
@@ -11,21 +11,21 @@ use Innmind\Immutable\Sequence;
 interface Node
 {
     /**
-     * @return Sequence<Node>
+     * @return Sequence<self|Element>
      */
     public function children(): Sequence;
 
     /**
-     * @param callable(Node): bool $filter
+     * @param callable(self|Element): bool $filter
      */
     public function filterChild(callable $filter): self;
 
     /**
-     * @param callable(Node): Node $map
+     * @param callable(self|Element): (self|Element) $map
      */
     public function mapChild(callable $map): self;
-    public function prependChild(self $child): self;
-    public function appendChild(self $child): self;
+    public function prependChild(self|Element $child): self;
+    public function appendChild(self|Element $child): self;
     public function content(): string;
     public function toString(): string;
 }