+++ src/Translator/NodeTranslator.php
@@ -3,7 +3,10 @@ declare(strict_types = 1);
 
 namespace Innmind\Xml\Translator;
 
-use Innmind\Xml\Node;
+use Innmind\Xml\{
+    Node,
+    Element,
+};
 use Innmind\Immutable\Maybe;
 
 /**
@@ -12,7 +15,7 @@ use Innmind\Immutable\Maybe;
 interface NodeTranslator
 {
     /**
-     * @return Maybe<Node>
+     * @return Maybe<Node|Element>
      */
     public function __invoke(\DOMNode $node, Translator $translate): Maybe;
 }