+++ src/Translator/Translator.php
@@ -3,7 +3,10 @@ declare(strict_types = 1);
 
 namespace Innmind\Xml\Translator;
 
-use Innmind\Xml\Node;
+use Innmind\Xml\{
+    Node,
+    Element,
+};
 use Innmind\Immutable\{
     Map,
     Maybe,
@@ -26,7 +29,7 @@ final class Translator
     }
 
     /**
-     * @return Maybe<Node>
+     * @return Maybe<Node|Element>
      */
     public function __invoke(\DOMNode $node): Maybe
     {