+++ src/Visitor/NextSibling.php
@@ -3,7 +3,10 @@ declare(strict_types = 1);
 
 namespace Innmind\Xml\Visitor;
 
-use Innmind\Xml\Node;
+use Innmind\Xml\{
+    Node,
+    Element,
+};
 use Innmind\Immutable\Maybe;
 
 /**
@@ -11,15 +14,15 @@ use Innmind\Immutable\Maybe;
  */
 final class NextSibling
 {
-    private Node $node;
+    private Node|Element $node;
 
-    private function __construct(Node $node)
+    private function __construct(Node|Element $node)
     {
         $this->node = $node;
     }
 
     /**
-     * @return Maybe<Node>
+     * @return Maybe<Node|Element>
      */
     public function __invoke(Node $tree): Maybe
     {
@@ -34,7 +37,7 @@ final class NextSibling
     /**
      * @psalm-pure
      */
-    public static function of(Node $node): self
+    public static function of(Node|Element $node): self
     {
         return new self($node);
     }