+++ src/Visitor/ParentNode.php
@@ -3,7 +3,10 @@ declare(strict_types = 1);
 
 namespace Innmind\Xml\Visitor;
 
-use Innmind\Xml\Node;
+use Innmind\Xml\{
+    Node,
+    Element,
+};
 use Innmind\Immutable\Maybe;
 
 /**
@@ -11,25 +14,25 @@ use Innmind\Immutable\Maybe;
  */
 final class ParentNode
 {
-    private Node $node;
+    private Node|Element $node;
 
-    private function __construct(Node $node)
+    private function __construct(Node|Element $node)
     {
         $this->node = $node;
     }
 
     /**
-     * @return Maybe<Node>
+     * @return Maybe<Node|Element>
      */
-    public function __invoke(Node $tree): Maybe
+    public function __invoke(Node|Element $tree): Maybe
     {
-        /** @var Maybe<Node> */
+        /** @var Maybe<Node|Element> */
         $parent = Maybe::nothing();
 
-        /** @var Maybe<Node> */
+        /** @var Maybe<Node|Element> */
         return $tree->children()->reduce(
             $parent,
-            function(Maybe $parent, Node $child) use ($tree): Maybe {
+            function(Maybe $parent, $child) use ($tree): Maybe {
                 if ($child === $this->node) {
                     return Maybe::just($tree);
                 }
@@ -42,7 +45,7 @@ final class ParentNode
     /**
      * @psalm-pure
      */
-    public static function of(Node $node): self
+    public static function of(Node|Element $node): self
     {
         return new self($node);
     }