+++ tests/Translator/NodeTranslator/ElementTranslatorTest.php
@@ -7,7 +7,6 @@ use Innmind\Xml\{
     Translator\NodeTranslator\ElementTranslator,
     Translator\NodeTranslator,
     Translator\Translator,
-    Node,
     Element,
     Element\Name,
 };
@@ -45,7 +44,7 @@ XML
                     new class($foo) implements NodeTranslator {
                         private $foo;
 
-                        public function __construct(Node $foo)
+                        public function __construct(Element $foo)
                         {
                             $this->foo = $foo;
                         }